+++ core/src/builder/script.rs
@@ -72,7 +72,7 @@ pub fn create_musig2_and_operator_multisig_script(
 /// < (`# of blocks mined after UTXO`) appears on the chain.
 pub fn generate_relative_timelock_script(
     actor_taproot_xonly_pk: XOnlyPublicKey, // This is the tweaked XonlyPublicKey, which appears in the script_pubkey of the address
-    block_count: u32,
+    block_count: i64,
 ) -> ScriptBuf {
     Builder::new()
         .push_int(block_count as i64)
